LayoutTests/imported/w3c/web-platform-tests/svg/geometry/parsing/x-invalid-expected.txt


FAIL e.style['x'] = "10" should not set the property value assert_equals: expected "" but got "10px"
PASS e.style['x'] = "auto" should not set the property value
PASS e.style['x'] = "10px 20px" should not set the property value
